+++ test/schemas.js
@@ -241,6 +241,10 @@ describe('Schemas', function() {
 
   this.timeout(t.timeout);
 
+  beforeEach(async function () {
+    apos.schema.validatedSchemas = {};
+  });
+
   before(async function() {
     apos = await t.create({
       root: module,
@@ -3129,6 +3133,43 @@ describe('Schemas', function() {
       assert(!output.requiredProp);
     });
 
+    // HERE
+    it('should not error nested required property if parent is not visible', async function() {
+      const req = apos.task.getReq();
+      const schema = apos.schema.compose({
+        addFields: [
+          {
+            name: 'object',
+            type: 'object',
+            if: {
+              showObject: true
+            },
+            schema: [
+              {
+                name: 'subfield',
+                type: 'string',
+                required: true
+              }
+            ]
+          },
+          {
+            name: 'showObject',
+            type: 'boolean'
+          }
+        ]
+      });
+      const output = {};
+
+      try {
+        await apos.schema.convert(req, schema, {
+          showObject: false
+        }, output);
+        assert(true);
+      } catch (err) {
+        assert(!err);
+      }
+    });
+
     it('should error required property nested boolean', async function() {
       const schema = apos.schema.compose({
         addFields: [
@@ -4946,7 +4987,484 @@ describe('Schemas', function() {
 
       await testSchemaError(schema, {}, 'age', 'required');
     });
+
+    it('should not error complex nested object required property if parents are not visible', async function() {
+      const schema = apos.schema.compose({
+        addFields: [
+          {
+            name: 'object',
+            type: 'object',
+            if: {
+              showObject: true
+            },
+            schema: [
+              {
+                name: 'objectString',
+                type: 'string',
+                required: true
+              },
+              {
+                name: 'objectArray',
+                type: 'array',
+                required: true,
+                if: {
+                  showObjectArray: true
+                },
+                schema: [
+                  {
+                    name: 'objectArrayString',
+                    type: 'string',
+                    required: true
+                  }
+                ]
+              },
+              {
+                name: 'showObjectArray',
+                type: 'boolean'
+              }
+            ]
+          },
+          {
+            name: 'showObject',
+            type: 'boolean'
+          }
+        ]
+      });
+
+      const data = {
+        object: {
+          objectString: 'toto',
+          objectArray: [
+            {
+              _id: 'tutu',
+              metaType: 'arrayItem'
+            }
+          ],
+          showObjectArray: false
+        },
+        showObject: true
+      };
+
+      const output = {};
+      const [ success, errors ] = await testConvert(apos, data, schema, output);
+
+      const expected = {
+        success: true,
+        errors: [],
+        output: {
+          object: {
+            _id: output.object._id,
+            objectString: 'toto',
+            objectArray: [
+              {
+                _id: 'tutu',
+                metaType: 'arrayItem',
+                scopedArrayName: undefined,
+                objectArrayString: ''
+              }
+            ],
+            showObjectArray: false,
+            metaType: 'objectItem',
+            scopedObjectName: undefined
+          },
+          showObject: true
+        }
+      };
+
+      const actual = {
+        success,
+        errors,
+        output
+      };
+
+      assert.deepEqual(expected, actual);
+
+    });
+
+    it('should not error complex nested arrays required property if parents are not visible', async function() {
+      const schema = apos.schema.compose({
+        addFields: [
+          {
+            name: 'root',
+            type: 'array',
+            if: {
+              showRoot: true
+            },
+            schema: [
+              {
+                name: 'rootString',
+                type: 'string',
+                required: true
+              },
+              {
+                name: 'rootArray',
+                type: 'array',
+                required: true,
+                schema: [
+                  {
+                    name: 'rootArrayString',
+                    type: 'string',
+                    required: true,
+                    if: {
+                      showRootArrayString: true
+                    }
+                  },
+                  {
+                    name: 'showRootArrayString',
+                    type: 'boolean'
+                  }
+                ]
+              }
+            ]
+          },
+          {
+            name: 'showRoot',
+            type: 'boolean'
+          }
+        ]
+      });
+
+      const data1 = {
+        root: [
+          {
+            _id: 'root_id',
+            metaType: 'arrayItem',
+            rootString: 'toto',
+            rootArray: [
+              {
+                _id: 'root_array_id',
+                metaType: 'arrayItem',
+                showRootArrayString: true
+              },
+              {
+                _id: 'root_array_id2',
+                metaType: 'arrayItem',
+                rootArrayBool: true,
+                showRootArrayString: false
+              }
+            ]
+          }
+        ],
+        showRoot: true
+      };
+
+      const output1 = {};
+      const [ success1, errors1 ] = await testConvert(apos, data1, schema, output1);
+      const foundError1 = findError(errors1, 'root_array_id.rootArrayString', 'required');
+
+      const data2 = {
+        root: [
+          {
+            _id: 'root_id',
+            metaType: 'arrayItem',
+            rootString: 'toto',
+            rootArray: [
+              {
+                _id: 'root_array_id',
+                metaType: 'arrayItem',
+                rootArrayString: 'Item 1',
+                showRootArrayString: true
+              },
+              {
+                _id: 'root_array_id2',
+                metaType: 'arrayItem',
+                rootArrayBool: true,
+                showRootArrayString: false
+              }
+            ]
+          }
+        ],
+        showRoot: true
+      };
+
+      const output2 = {};
+      const [ success2, errors2 ] = await testConvert(apos, data2, schema, output2);
+
+      const expected = {
+        success1: false,
+        foundError1: true,
+        success2: true,
+        errors2: [],
+        output2: {
+          root: [
+            {
+              _id: 'root_id',
+              metaType: 'arrayItem',
+              scopedArrayName: undefined,
+              rootString: 'toto',
+              rootArray: [
+                {
+                  _id: 'root_array_id',
+                  metaType: 'arrayItem',
+                  scopedArrayName: undefined,
+                  rootArrayString: 'Item 1',
+                  showRootArrayString: true
+                },
+                {
+                  _id: 'root_array_id2',
+                  metaType: 'arrayItem',
+                  scopedArrayName: undefined,
+                  rootArrayString: '',
+                  showRootArrayString: false
+                }
+              ]
+            }
+          ],
+          showRoot: true
+        }
+      };
+
+      const actual = {
+        success1,
+        foundError1,
+        success2,
+        errors2,
+        output2
+      };
+
+      assert.deepEqual(expected, actual);
+    });
+
+    // TODO: update this test when support for conditional fields is added to relationships schemas
+    it('should not error complex nested relationships required property if parents are not visible', async function() {
+      const req = apos.task.getReq({ mode: 'draft' });
+      const schema = apos.schema.compose({
+        addFields: [
+          {
+            name: 'title',
+            type: 'string',
+            required: true
+          },
+          {
+            name: '_rel',
+            type: 'relationship',
+            withType: 'article',
+            schema: [
+              {
+                name: 'relString',
+                type: 'string',
+                required: true,
+                if: {
+                  showRelString: true
+                }
+              },
+              {
+                name: 'showRelString',
+                type: 'boolean'
+              }
+            ]
+          }
+        ]
+      });
+
+      const article1 = await apos.article.insert(req, {
+        ...apos.article.newInstance(),
+        title: 'article 1'
+      });
+      const article2 = await apos.article.insert(req, {
+        ...apos.article.newInstance(),
+        title: 'article 2'
+      });
+
+      article1._fields = {
+        showRelString: false
+      };
+
+      article2._fields = {
+        relString: 'article 2 rel string',
+        showRelString: true
+      };
+
+      const data = {
+        title: 'toto',
+        _rel: [
+          article1,
+          article2
+        ]
+      };
+
+      const output = {};
+      const [ success, errors ] = await testConvert(apos, data, schema, output);
+      const foundError = findError(errors, 'relString', 'required');
+
+      const expected = {
+        success: false,
+        foundError: true
+      };
+
+      const actual = {
+        success,
+        foundError
+      };
+
+      assert.deepEqual(expected, actual);
+    });
+
+    it('should add proper aposPath to all fields when validation schema', async function () {
+      const schema = apos.schema.compose({
+        addFields: [
+          {
+            name: 'title',
+            type: 'string',
+            required: true
+          },
+          {
+            name: '_rel',
+            type: 'relationship',
+            withType: 'article',
+            schema: [
+              {
+                name: 'relString',
+                type: 'string'
+              }
+            ]
+          },
+          {
+            name: 'array',
+            type: 'array',
+            schema: [
+              {
+                name: 'arrayString',
+                type: 'string'
+              },
+              {
+                name: 'arrayObject',
+                type: 'object',
+                schema: [
+                  {
+                    name: 'arrayObjectString',
+                    type: 'string'
+                  }
+                ]
+              }
+            ]
+          },
+          {
+            name: 'object',
+            type: 'object',
+            schema: [
+              {
+                name: 'objectString',
+                type: 'string'
+              },
+              {
+                name: 'objectArray',
+                type: 'array',
+                schema: [
+                  {
+                    name: 'objectArrayString',
+                    type: 'string'
+                  }
+                ]
+              }
+            ]
+          }
+        ]
+      });
+
+      apos.schema.validate(schema, 'article');
+
+      const [ titleField, relField, arrayField, objectField ] = schema;
+      const expected = {
+        title: 'title',
+        rel: '_rel',
+        relString: '_rel/relString',
+        array: 'array',
+        arrayString: 'array/arrayString',
+        arrayObject: 'array/arrayObject',
+        arrayObjectString: 'array/arrayObject/arrayObjectString',
+        object: 'object',
+        objectString: 'object/objectString',
+        objectArray: 'object/objectArray',
+        objectArrayString: 'object/objectArray/objectArrayString'
+      };
+
+      const actual = {
+        title: titleField.aposPath,
+        rel: relField.aposPath,
+        relString: relField.schema[0].aposPath,
+        array: arrayField.aposPath,
+        arrayString: arrayField.schema[0].aposPath,
+        arrayObject: arrayField.schema[1].aposPath,
+        arrayObjectString: arrayField.schema[1].schema[0].aposPath,
+        object: objectField.aposPath,
+        objectString: objectField.schema[0].aposPath,
+        objectArray: objectField.schema[1].aposPath,
+        objectArrayString: objectField.schema[1].schema[0].aposPath
+      };
+
+      assert.deepEqual(actual, expected);
+    });
+
+    it('should set default value to invisible fields that triggered convert errors', async function () {
+      const schema = apos.schema.compose({
+        addFields: [
+          {
+            name: 'array',
+            type: 'array',
+            if: {
+              showArray: true
+            },
+            schema: [
+              {
+                name: 'arrayString',
+                type: 'string',
+                pattern: '^cool-'
+              },
+              {
+                name: 'arrayMin',
+                type: 'integer',
+                min: 5
+              },
+              {
+                name: 'arrayMax',
+                type: 'integer',
+                max: 10
+              }
+            ]
+          },
+          {
+            name: 'showArray',
+            type: 'boolean'
+          }
+        ]
+      });
+      apos.schema.validate(schema, 'foo');
+
+      const input = {
+        showArray: false,
+        array: [
+          {
+            arrayString: 'bad string',
+            arrayMin: 2,
+            arrayMax: 13
+          }
+        ]
+      };
+
+      const req = apos.task.getReq();
+      const result = {};
+      await apos.schema.convert(req, schema, input, result);
+
+      const expected = {
+        arrayString: '',
+        arrayMin: 5,
+        arrayMax: 10
+      };
+
+      const {
+        arrayString, arrayMin, arrayMax
+      } = result.array[0];
+      const actual = {
+        arrayString,
+        arrayMin,
+        arrayMax
+      };
+
+      assert.deepEqual(actual, expected);
+    });
   });
+
   async function testSchemaError(schema, input, path, name) {
     const req = apos.task.getReq();
     const result = {};
@@ -4965,3 +5483,34 @@ describe('Schemas', function() {
     }
   }
 });
+
+async function testConvert(
+  apos,
+  data,
+  schema,
+  output
+) {
+  const req = apos.task.getReq({ mode: 'draft' });
+  try {
+    await apos.schema.convert(req, schema, data, output);
+    return [ true, [] ];
+  } catch (err) {
+    return [ false, err ];
+  }
+}
+
+function findError(errors, fieldPath, errorName) {
+  if (!Array.isArray(errors)) {
+    return false;
+  }
+  return errors.some((err) => {
+    if (err.data?.errors) {
+      const deepFound = findError(err.data.errors, fieldPath, errorName);
+      if (deepFound) {
+        return deepFound;
+      }
+    }
+
+    return err.name === errorName && err.path === fieldPath;
+  });
+}